+++ src/Utility/BackupManager.php
@@ -21,9 +21,9 @@ use Cake\I18n\FrozenTime;
 use Cake\Mailer\Mailer;
 use Cake\ORM\Entity;
 use DatabaseBackup\BackupTrait;
+use LogicException;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Finder\SplFileInfo;
-use Tools\Exceptionist;
 use Tools\Filesystem;
 
 /**
@@ -38,11 +38,14 @@ class BackupManager
      * @param string $filename Backup filename you want to delete. The path can be relative to the backup directory
      * @return string Deleted backup file
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupManager-utility#delete
-     * @throws \Tools\Exception\NotWritableException
+     * @throws \LogicException
      */
     public static function delete(string $filename): string
     {
-        $filename = Exceptionist::isWritable(self::getAbsolutePath($filename));
+        $filename = self::getAbsolutePath($filename);
+        if (!is_writable($filename)) {
+            throw new LogicException(__d('database_backup', 'File or directory `' . $filename . '` is not writable'));
+        }
         Filesystem::instance()->remove($filename);
 
         return $filename;
@@ -87,12 +90,13 @@ class BackupManager
      * @param int $rotate Number of backups that you want to keep
      * @return array<\Cake\ORM\Entity> Array of deleted files
      * @see https://github.com/mirko-pagliai/cakephp-database-backup/wiki/How-to-use-the-BackupManager-utility#rotate
-     * @throws \ErrorException
-     * @noinspection PhpDocRedundantThrowsInspection
+     * @throws \LogicException
      */
     public static function rotate(int $rotate): array
     {
-        Exceptionist::isPositive($rotate, __d('database_backup', 'Invalid rotate value'));
+        if (!is_positive($rotate)) {
+            throw new LogicException(__d('database_backup', 'Invalid rotate value'));
+        }
         $backupsToBeDeleted = self::index()->skip($rotate);
         array_map([__CLASS__, 'delete'], $backupsToBeDeleted->extract('filename')->toList());
 
@@ -105,18 +109,21 @@ class BackupManager
      * @param string $recipient Recipient's email address
      * @return \Cake\Mailer\Mailer
      * @since 1.1.0
-     * @throws \Tools\Exception\NotReadableException
+     * @throws \LogicException
      */
     protected static function getEmailInstance(string $backup, string $recipient): Mailer
     {
-        $file = Exceptionist::isReadable(self::getAbsolutePath($backup));
+        $filename = self::getAbsolutePath($backup);
+        if (!is_readable($filename)) {
+            throw new LogicException(__d('database_backup', 'File or directory `' . $filename . '` is not readable'));
+        }
         $server = env('SERVER_NAME', 'localhost');
 
         return (new Mailer())
             ->setFrom(Configure::readOrFail('DatabaseBackup.mailSender'))
             ->setTo($recipient)
-            ->setSubject(__d('database_backup', 'Database backup {0} from {1}', basename($file), $server))
-            ->setAttachments([basename($file) => compact('file') + ['mimetype' => mime_content_type($file)]]);
+            ->setSubject(__d('database_backup', 'Database backup {0} from {1}', basename($filename), $server))
+            ->setAttachments([basename($filename) => ['file' => $filename, 'mimetype' => mime_content_type($filename)]]);
     }
 
     /**
@@ -124,7 +131,7 @@ class BackupManager
      * @param string $filename Backup filename you want to send via email. The path can be relative to the backup directory
      * @param string $recipient Recipient's email address
      * @return array{headers: string, message: string}
-     * @throws \Tools\Exception\NotReadableException
+     * @throws \LogicException
      * @since 1.1.0
      */
     public static function send(string $filename, string $recipient): array